requirements.txt
# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# API & AI/ML Libraries
openai==1.3.7
google-generativeai==0.3.2
httpx==0.25.2

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0

# Monitoring & Logging
wandb==0.17.0
psutil==5.9.6
sentry-sdk==1.38.0

# Image Processing
Pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0

# Other Utilities
requests==2.31.0
aiofiles==23.2.1
